Written a new LDA parser which parses the stowed LDAs instead of the source code
