--- libs/nat-lda-parser/src/old/test.c
@@ -1,14 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#include "../include/standard.h"
-#include "parser.h"
-
-int main()
-{
-    struct varnames *vars;
-    
-    vars = getVarNames("../../TESTLDA.NSL");
-    printfork(vars);
-
-}
